feat: added functionality to check if the datasets folder exists if it does not create it

+++ api/repository/functionality.py
@@ -1,6 +1,7 @@
 from fastapi import UploadFile
 from fastapi.responses import JSONResponse
 from pathlib import Path
+import os
 import logging
 import logging.config
 
@@ -24,21 +25,33 @@ def uploadFile(file: UploadFile) -> JSONResponse:
         JSONResponse: {'saved':Boolean,'path':filepath uploaded}
     """
     try:
+        # Ask if database folder had been creted if not create it and save file
+        if Path(f"{rootPath}/datasets").exists():
+            with open(f"{rootPath}/datasets/{file.filename}", 'wb') as myFile:
+                # With second file name convert the first in bynary
+                # and then read the binary
+                content = file.file.read()
+                myFile.write(content)
+                myFile.close()
+                logger.info('File successfuly created')
+            return JSONResponse(content={
+                'saved': True,
+                'path': f'{rootPath}/{file.filename}'
+            }, status_code=200)
+        else:
+            os.mkdir(f'{rootPath}/datasets')
+            with open(f"{rootPath}/datasets/{file.filename}", 'wb') as myFile:
+                content = file.file.read()
+                myFile.write(content)
+                myFile.close()
+                logger.info('File successfuly created')
+            return JSONResponse(content={
+                'saved': True,
+                'path': f'{rootPath}/{file.filename}'
+            }, status_code=200)
 
-        with open(f"{rootPath}/datasets/{file.filename}", 'wb') as myFile:
-            # With second file name convert the first in bynary
-            # and then read the binary
-            content = file.file.read()
-            myFile.write(content)
-            myFile.close()
-            logger.info('File successfuly created')
-        return JSONResponse(content={
-            'saved': True,
-            'path': f'{rootPath}/{file.filename}'
-        }, status_code=200)
-
-    except FileNotFoundError:
-        logger.error(f'Error File not found: {FileNotFoundError}')
+    except FileNotFoundError:        
+        logger.error(f'Error File not found, directory had been created: {FileNotFoundError}')
         return JSONResponse(content={
             'saved': False
         }, status_code=404)
